Move closed channels tracking to LightningRepo
