+++ app/src/main/java/to/bitkit/services/LightningService.kt
@@ -1,7 +1,5 @@
 package to.bitkit.services
 
-import com.synonym.bitkitcore.ClosedChannelDetails
-import com.synonym.bitkitcore.upsertClosedChannel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.currentCoroutineContext
 import kotlinx.coroutines.delay
@@ -51,7 +49,6 @@ import to.bitkit.utils.LdkError
 import to.bitkit.utils.LdkLogWriter
 import to.bitkit.utils.Logger
 import to.bitkit.utils.ServiceError
-import java.util.concurrent.ConcurrentHashMap
 import javax.inject.Inject
 import javax.inject.Singleton
 import kotlin.io.path.Path
@@ -73,8 +70,6 @@ class LightningService @Inject constructor(
     var node: Node? = null
 
     private lateinit var trustedPeers: List<PeerDetails>
-
-    private val channelCache = ConcurrentHashMap<String, ChannelDetails>()
 
     suspend fun setup(
         walletIndex: Int,
@@ -196,7 +191,6 @@ class LightningService @Inject constructor(
             }
         }
 
-        refreshChannelCache()
         Logger.info("Node started")
     }
 
@@ -232,77 +226,8 @@ class LightningService @Inject constructor(
             node.syncWallets()
             // launch { setMaxDustHtlcExposureForCurrentChannels() }
         }
-        refreshChannelCache()
 
         Logger.debug("LDK synced")
-    }
-
-    private suspend fun refreshChannelCache() {
-        val node = this.node ?: return
-
-        ServiceQueue.LDK.background {
-            val channels = node.listChannels()
-            channels.forEach { channel ->
-                channelCache[channel.channelId] = channel
-            }
-        }
-    }
-
-    private suspend fun registerClosedChannel(channelId: String, reason: String?) {
-        try {
-            val channel = ServiceQueue.LDK.background {
-                channelCache[channelId]
-            } ?: run {
-                Logger.error(
-                    "Could not find channel details for closed channel: channelId=$channelId",
-                    context = TAG
-                )
-                return@registerClosedChannel
-            }
-
-            val fundingTxo = channel.fundingTxo
-            if (fundingTxo == null) {
-                Logger.error(
-                    "Channel has no funding transaction, cannot persist closed channel: channelId=$channelId",
-                    context = TAG
-                )
-                return@registerClosedChannel
-            }
-
-            val channelName = channel.inboundScidAlias?.toString()
-                ?: channel.channelId.take(CHANNEL_ID_PREVIEW_LENGTH) + "…"
-
-            val closedAt = (System.currentTimeMillis() / 1000L).toULong()
-
-            val closedChannel = ClosedChannelDetails(
-                channelId = channel.channelId,
-                counterpartyNodeId = channel.counterpartyNodeId,
-                fundingTxoTxid = fundingTxo.txid,
-                fundingTxoIndex = fundingTxo.vout,
-                channelValueSats = channel.channelValueSats,
-                closedAt = closedAt,
-                outboundCapacityMsat = channel.outboundCapacityMsat,
-                inboundCapacityMsat = channel.inboundCapacityMsat,
-                counterpartyUnspendablePunishmentReserve = channel.counterpartyUnspendablePunishmentReserve,
-                unspendablePunishmentReserve = channel.unspendablePunishmentReserve ?: 0u,
-                forwardingFeeProportionalMillionths = channel.config.forwardingFeeProportionalMillionths,
-                forwardingFeeBaseMsat = channel.config.forwardingFeeBaseMsat,
-                channelName = channelName,
-                channelClosureReason = reason.orEmpty()
-            )
-
-            ServiceQueue.CORE.background {
-                upsertClosedChannel(closedChannel)
-            }
-
-            ServiceQueue.LDK.background {
-                channelCache.remove(channelId)
-            }
-
-            Logger.info("Registered closed channel: ${channel.userChannelId}", context = TAG)
-        } catch (e: Exception) {
-            Logger.error("Failed to register closed channel: $e", e, context = TAG)
-        }
     }
 
     // private fun setMaxDustHtlcExposureForCurrentChannels() {
@@ -815,9 +740,6 @@ class LightningService @Inject constructor(
                 Logger.info(
                     "⏳ Channel pending: channelId: $channelId userChannelId: $userChannelId formerTemporaryChannelId: $formerTemporaryChannelId counterpartyNodeId: $counterpartyNodeId fundingTxo: $fundingTxo"
                 )
-                launch {
-                    refreshChannelCache()
-                }
             }
 
             is Event.ChannelReady -> {
@@ -827,9 +749,6 @@ class LightningService @Inject constructor(
                 Logger.info(
                     "👐 Channel ready: channelId: $channelId userChannelId: $userChannelId counterpartyNodeId: $counterpartyNodeId"
                 )
-                launch {
-                    refreshChannelCache()
-                }
             }
 
             is Event.ChannelClosed -> {
@@ -840,9 +759,6 @@ class LightningService @Inject constructor(
                 Logger.info(
                     "⛔ Channel closed: channelId: $channelId userChannelId: $userChannelId counterpartyNodeId: $counterpartyNodeId reason: $reason"
                 )
-                launch {
-                    registerClosedChannel(channelId, reason)
-                }
             }
         }
     }
@@ -867,7 +783,6 @@ class LightningService @Inject constructor(
 
     companion object {
         private const val TAG = "LightningService"
-        private const val CHANNEL_ID_PREVIEW_LENGTH = 10
     }
 }
